fix(db): cast has_org_role rank literals to smallint in 072
A dry-run of 072 inside BEGIN ... ROLLBACK failed with:

  ERROR: 42883: function public.has_org_role(uuid, integer) does not exist
  HINT:  No function matches the given name and argument types.

has_org_role(target_org uuid, min_rank smallint) takes smallint, and an
uncast integer literal is not implicitly matched - which is why the live
policies read has_org_role(organization_id, (2)::smallint). All rank
literals are now cast explicitly and the migration dry-runs cleanly.

Recorded because it is exactly the class of failure that "the file exists,
therefore it is applied" reasoning hides.

+++ database/migrations/072_harden_organization_and_member_rls.sql
@@ -70,8 +70,8 @@ drop policy if exists org_update on public.organizations;
 create policy org_update on public.organizations
   for update
   to authenticated
-  using (public.has_org_role(id, 2))
-  with check (public.has_org_role(id, 2));
+  using (public.has_org_role(id, 2::smallint))
+  with check (public.has_org_role(id, 2::smallint));
 
 -- ---------------------------------------------------------------------
 -- FINDING 3/4/5 — membership writes require OWNER or ADMIN.
@@ -88,21 +88,21 @@ drop policy if exists members_insert on public.organization_members;
 create policy members_insert on public.organization_members
   for insert
   to authenticated
-  with check (public.has_org_role(organization_id, 2));
+  with check (public.has_org_role(organization_id, 2::smallint));
 
 drop policy if exists members_update on public.organization_members;
 create policy members_update on public.organization_members
   for update
   to authenticated
-  using (public.has_org_role(organization_id, 2))
+  using (public.has_org_role(organization_id, 2::smallint))
   with check (
-    public.has_org_role(organization_id, 2)
+    public.has_org_role(organization_id, 2::smallint)
     and (
       -- Granting OWNER requires being an OWNER.
       role_id <> (
         select r.id from public.organization_roles r where r.code = 'OWNER'
       )
-      or public.has_org_role(organization_id, 1)
+      or public.has_org_role(organization_id, 1::smallint)
     )
   );
 
@@ -111,7 +111,7 @@ create policy members_delete on public.organization_members
   for delete
   to authenticated
   using (
-    public.has_org_role(organization_id, 2)
+    public.has_org_role(organization_id, 2::smallint)
     and (
       -- The OWNER membership is not removable through the client API.
       role_id <> (
